refactor: location endpoint.
