+++ src/endpoints/location/models/state.entity.ts
