+++ src/substrate/substrate.module.ts
@@ -5,7 +5,7 @@ import { EscrowModule } from '../escrow/escrow.module';
 import { TransactionLoggingModule } from '../transaction-logging/transaction-logging.module';
 import { RewardModule } from 'src/reward/reward.module';
 import { MailModule } from 'src/common/mailer/mailer.module';
-import { LocationModule } from 'src/location/location.module';
+import { LocationModule } from 'src/endpoints/location/location.module';
 import { DebioConversionModule } from 'src/debio-conversion/debio-conversion.module';
 
 @Module({
