+++ test/unit/location/country.service.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { getRepositoryToken } from '@nestjs/typeorm';
-import { CountryService } from '../../../src/location/country.service';
-import { Country } from '../../../src/location/models/country.entity';
+import { CountryService } from '../../../src/endpoints/location/country.service';
+import { Country } from '../../../src/endpoints/location/models/country.entity';
 import { MockType, repositoryMockFactory } from '../mock';
 import { Repository } from 'typeorm';
 import { when } from 'jest-when';
